requirements.txt
asgiref==3.9.1
Django==5.2.6
tzdata==2025.2
whitenoise==6.9.0
sqlparse==0.5.3
django-colorfield==0.14.0
pillow==11.3.0
